backend/requirements.txt
deepgram-sdk==2.12.0
feedparser
python-dotenv
fastapi
uvicorn
pydantic
elevenlabs